feat: confirm password & show/hide password eye

+++ src/frontend/contexts/AuthProvider.js
@@ -27,6 +27,7 @@ export default function AuthProvider({ children }) {
     lastName: "",
     email: "",
     password: "",
+    confirmPassword: "",
   });
   const [userLoginData, setUserLoginData] = useState({
     email: "adarshbalika@gmail.com",
@@ -34,29 +35,47 @@ export default function AuthProvider({ children }) {
   });
 
   const signupHandler = async () => {
-    try {
-      const {
-        data: { createdUser, encodedToken },
-      } = await axios.post("/api/auth/signup", userSignupData);
-      const user = {
-        encodedToken: encodedToken,
-        firstName: createdUser.firstName.trim(),
-        lastName: createdUser.firstName.trim(),
-        email: createdUser.email.trim(),
-      };
-      setCurrentUser(user);
-      localStorage.setItem("token", encodedToken);
-      localStorage.setItem("user", JSON.stringify(user));
-      setLoggedIn(true);
-      toast.success("You're signedUp!", {
-        className: "toast-message",
-      });
-      fromLocation === undefined
-        ? navigate("/")
-        : navigate(location.state?.from?.pathname);
-    } catch (e) {
-      console.log(e.response.data.errors[0]);
-      toast.error(e.response.data.errors[0], {
+    if (
+      userSignupData.firstName &&
+      userSignupData.lastName &&
+      userSignupData.email &&
+      userSignupData.password &&
+      userSignupData.confirmPassword
+    ) {
+      if (userSignupData.password === userSignupData.confirmPassword) {
+        try {
+          const {
+            data: { createdUser, encodedToken },
+          } = await axios.post("/api/auth/signup", userSignupData);
+          const user = {
+            encodedToken: encodedToken,
+            firstName: createdUser.firstName.trim(),
+            lastName: createdUser.firstName.trim(),
+            email: createdUser.email.trim(),
+          };
+          setCurrentUser(user);
+          localStorage.setItem("token", encodedToken);
+          localStorage.setItem("user", JSON.stringify(user));
+          setLoggedIn(true);
+          toast.success("You're signedUp!", {
+            className: "toast-message",
+          });
+          fromLocation === undefined
+            ? navigate("/")
+            : navigate(location.state?.from?.pathname);
+        } catch (e) {
+          console.log(e.response.data.errors[0]);
+          toast.error(e.response.data.errors[0], {
+            className: "toast-message",
+          });
+        }
+      } else {
+        toast.warning("Password and confirm password doesn't match!!", {
+          className: "toast-message",
+        });
+      }
+    } else {
+      toast.warning("Fill all the input requirements!!", {
         className: "toast-message",
       });
     }
